backend/requirements.txt
# backend/requirements.txt（修正版）

# FastAPI本体 + サーバー
fastapi==0.110.1
uvicorn[standard]==0.29.0

# データベース関連
sqlalchemy==2.0.29

# モデル/スキーマ定義
pydantic==2.9.0

# LLM推論（llama.cpp統合）
llama-cpp-python==0.2.90

# ワードクラウド関連
wordcloud==1.9.3
janome==0.5.0
pillow==10.3.0

# システム情報取得
cpuinfo==9.0.0
psutil==5.9.6

# .env対応（互換性のため残す）
python-dotenv==1.0.1

# 開発・テスト用ツール
ruff>=0.4.4,<1.0.0
pytest==8.2.0
httpx==0.27.0

# ビルド用
pyinstaller==6.3.0

# Hugging Face（モデルダウンロード用）
huggingface-hub==0.20.0
